Revert tabs changes and restructure tabs
